third_party/WebKit/LayoutTests/fast/frames/iframe-no-name.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/the-iframe-element.html#attr-iframe-name">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<iframe id="foo"></iframe>
<script>
description("Checks that the iframe's name is an empty string if unset.");
shouldBeEqualToString("frames[0].name", "");
</script>
</body>
</html>
